# services/web-portal/pkg/infrastructure/rest/controllers/invoice/responses.go
// Package invoice contains the invoice controller
package invoice

// MessageResponse is a struct that contains the response body for the message
type MessageResponse struct {
	Message string `json:"message"`
}